# rose-backend-main/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/user/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/user/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/user/set-admin/{id}": {
      "post": {
        "operationId": "AuthController_setUserAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/user/last-login": {
      "get": {
        "operationId": "AuthController_lastLoginDate",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/image": {
      "post": {
        "operationId": "ImageController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "alt": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "name",
                  "image"
                ]
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "alt": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "name",
                  "image"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "images"
        ]
      },
      "get": {
        "operationId": "ImageController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "images"
        ]
      }
    },
    "/image/{id}": {
      "get": {
        "operationId": "ImageController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "images"
        ]
      },
      "patch": {
        "operationId": "ImageController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "alt": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "alt": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "images"
        ]
      },
      "delete": {
        "operationId": "ImageController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "images"
        ]
      }
    },
...
    "/menu/items": {
      "post": {
        "operationId": "MenuController_createMenuItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuItemDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Menu"
        ]
      },
      "get": {
        "operationId": "MenuController_getMenuItems",
        "parameters": [
          {
            "name": "restaurantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "example": "1",
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "string",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/menu/items/{id}": {
      "get": {
        "operationId": "MenuController_findMenuItemById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Menu"
        ]
      },
      "patch": {
        "operationId": "MenuController_updateMenuItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Menu"
        ]
      },
      "delete": {
        "operationId": "MenuController_removeMenuItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Menu"
        ]
      }
    },
    "/menu/categories": {
      "post": {
        "operationId": "MenuController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Menu"
        ]
      },
      "get": {
        "operationId": "MenuController_getAllCategories",
        "parameters": [
          {
            "name": "restaurantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/menu/categories/{id}": {
      "get": {
        "operationId": "MenuController_findCategoryById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Menu"
        ]
      },
      "patch": {
        "operationId": "MenuController_updateCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Menu"
        ]
      },
      "delete": {
        "operationId": "MenuController_removeCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Menu"
        ]
      }
    },
    "/menu/items/category/{id}": {
      "get": {
        "operationId": "MenuController_getItemsByCategoryId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
...
  },
  "info": {
    "title": "YouTaab API",
    "description": "API for coffeshop reservation and menu",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "fullname": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirm_password": {
            "type": "string"
          }
        },
        "required": [
          "fullname",
          "username",
          "password",
          "confirm_password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "CreateMenuItemDto": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fee": {
            "type": "number"
          },
          "available": {
            "type": "boolean"
          },
          "categoryId": {
            "type": "number"
          }
        },
        "required": [
          "title",
          "fee",
          "categoryId"
        ]
      },
      "UpdateMenuItemDto": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fee": {
            "type": "number"
          },
          "available": {
            "type": "boolean"
          },
          "categoryId": {
            "type": "number"
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Appetizers"
          },
          "icon": {
            "type": "string",
            "example": "utensils"
          },
          "restaurantId": {
            "type": "number",
            "example": 12,
            "description": "شناسه رستورانِ مربوط به این دسته‌بندی"
          }
        },
        "required": [
          "name",
          "icon",
          "restaurantId"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Appetizers"
          },
          "icon": {
            "type": "string",
            "example": "utensils"
          }
        }
      },
...
    }
  }
}